User Requirments
Functional Requirements

The Functional Requirement is the ability of a user to interact with the software application using reliable user interface design (Ofni Systems, 2017). 
o	Login/register- new engineers have to register themselves into the application to access the services. The existing users have to enter their valid email address and password to continue accessing with the services. 
o	De-register an engineer by the admin - This operation is required to eliminate the details of an engineer who no longer works with the company or was sucked. Or it can be used to eliminate any wrong entry in the system. 
o	Create job- There has to be an option ‘new’ to create a new entry in the system for a specified engineer. Second option should be to update the details of a particular engineer. 
o	Customer Feedback- customers have to login first to provide their valuable feedback for the services accepted from the company. 
o	Add a specialization to an engineer when he is registering
o	Approve a job by the admin when it is created by a customer 
o	Authorise or de-authorise an engineer by the admin when the engineer registered
o	Cancel an on progress or created job by a customer or by an engineer
o	Generate weekly reports of jobs with all the afferent details of it
o	Select job by an engineer when the job is not assigned to any engineer by the customer
o	Update job status by an engineer 
o	View the job status by a customer

Non-Functional Requirements

o	Performance- the response time should be low so that users can access the system fatly. This affects the usability of the desktop application because more the response time will, less the customers will be interested to use the services. 
o	Reliability- the system has to be consistent in performance; it means that there should be less or no software failure.
o	Security- authorization techniques should be used to ensure that the stored data in the system is protected from any kind of malicious activity. 
o	Usability- there should be no need of users to make any efforts in accessing the system. The customer service system 
o	Portability- system should have the capability to be accessed on different operating systems. 
o	Modifiability- the system should be flexible enough so that relevant changes can be made according to the demands of customers in future